added compute shader for marching cubes

+++ Source/GeoFlowCore/Private/GeoFlowConstantNodes.cpp
@@ -20,7 +20,7 @@ UGFN_R_Base* UGFN_E_ConstantInt::CreateRuntimeNode(UGeoFlowRuntimeGraph* runtime
 UGFN_R_Base* UGFN_E_ConstantDouble::CreateRuntimeNode(UGeoFlowRuntimeGraph* runtimeGraph, TArray<std::pair<FGuid, FGuid>>& connections, TMap<FGuid, UGeoFlowRuntimePin*>& idToPinMap)
 {
 	UGFN_R_ConstantDouble* runtimeNode = InitRuntimeNode<UGFN_R_ConstantDouble>(runtimeGraph);
-	runtimeNode->Output = InitRuntimePin(runtimeNode, Output, connections, idToPinMap, EGeoFlowReturnType::Double);
+	runtimeNode->Output = InitRuntimePin(runtimeNode, Output, connections, idToPinMap, EGeoFlowReturnType::Float);
 	runtimeNode->Value = Value;
 	return runtimeNode;
 }
@@ -41,6 +41,7 @@ UGFN_E_Base* UGFN_R_ConstantBool::CreateEditorNode(UEdGraph* _workingGraph, TArr
 	newNode->Value = Value;
 	return newNode;
 }
+
 UGFN_E_Base* UGFN_R_ConstantInt::CreateEditorNode(UEdGraph* _workingGraph, TArray<std::pair<FGuid, FGuid>>& connections, TMap<FGuid, UEdGraphPin*>& idToPinMap)
 {
 	UGFN_E_ConstantInt* newNode = InitUiNode<UGFN_E_ConstantInt>(_workingGraph);
@@ -48,6 +49,7 @@ UGFN_E_Base* UGFN_R_ConstantInt::CreateEditorNode(UEdGraph* _workingGraph, TArra
 	newNode->Value = Value;
 	return newNode;
 }
+
 UGFN_E_Base* UGFN_R_ConstantDouble::CreateEditorNode(UEdGraph* _workingGraph, TArray<std::pair<FGuid, FGuid>>& connections, TMap<FGuid, UEdGraphPin*>& idToPinMap)
 {
 	UGFN_E_ConstantDouble* newNode = InitUiNode<UGFN_E_ConstantDouble>(_workingGraph);
@@ -55,6 +57,7 @@ UGFN_E_Base* UGFN_R_ConstantDouble::CreateEditorNode(UEdGraph* _workingGraph, TA
 	newNode->Value = Value;
 	return newNode;
 }
+
 UGFN_E_Base* UGFN_R_ConstantVector::CreateEditorNode(UEdGraph* _workingGraph, TArray<std::pair<FGuid, FGuid>>& connections, TMap<FGuid, UEdGraphPin*>& idToPinMap)
 {
 	UGFN_E_ConstantVector* newNode = InitUiNode<UGFN_E_ConstantVector>(_workingGraph);
@@ -62,3 +65,4 @@ UGFN_E_Base* UGFN_R_ConstantVector::CreateEditorNode(UEdGraph* _workingGraph, TA
 	newNode->Value = Value;
 	return newNode;
 }
+
